Move special offsets from ExchangeCalendar to XKRX
Moves 'special offsets' implementation from ExchangeCalendar to XKRX.
Leaves a `apply_special_offsets` hook into the ExchangeCalendar
constructor.
